LayoutTests/accessibility/scroll-to-global-point-iframe-expected.txt

Tests that scrolling to move an element to a specific point successfully scrolls an iframe.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


5000-pixel box

5000-pixel box
PASS window.pageYOffset is 0
PASS frameWindow.pageYOffset is 0
PASS target.getBoundingClientRect().top is 0
PASS target.getBoundingClientRect().top is 300
PASS target.getBoundingClientRect().top is 3000
PASS successfullyParsed is true

TEST COMPLETE
